Share test help code (#6)
* Update Readme

- Tests and linting added in [#4, #5]
- Err handling introduced in #3

* Whitespace

* Share test help code

Previously the build and detect modules had a lot of duplicate logic around creating a context for testing. This commit moves the logic into their own file so it can be shared.

The ultimate goal of this exercise would be to work towards a generic context factory we can put into libcnb so other build packs don't have to re-implement the same logic.

The intermediate goal is to explore patterns for internal code-reuse for the purposes of testing.