Rename BuildpackError enum to ProcfileError (#8)
It was confusing to me, while learning, that `libcnb::Error` has a `libcnb::Error::BuildpackError` and also that the profile buildpack has a BuildpackError. I intentionally want to rename it something different so the error handling is a little clearer. 

I intend this to be a "teaching" build pack that can be used for reference implementations. As such, I've extensively documented how the errors are structured and given an example on how to work from "I got an error because we couldn't convert between errors" to a working buildpack.