Remove unnecessary install of `libcnb-cargo` in CI (#47)
Since it's not required to run the integration tests, and will only used by
the release process, which will happen separately in a GitHub Action as
part of #41. 

This saves ~4 minutes of CI time when the cache is cold (which is any time
`Cargo.lock` or the Circle CI config changes), or when a new version of
`libcnb-cargo` has been released.

GUS-W-10736354.