Fix posting a release to GitHub
The prior version of this release script used the tag only which means that it's required that a PR have the version number as the git ref (branch name). I want to decouple this and we already have the version number.

Heroku Ruby uses the `tag_name` to specify a custom value, we want to set it to the version https://github.com/heroku/heroku-buildpack-ruby/blob/d42226c1dc0dfbc123b1facaeae276cf3d2689a6/.github/workflows/release.yml#L43-L55.